Extract notifications into separate component
